+++ oscar-cbls/src/main/examples/oscar/examples/cbls/routing/SimpleVRPWithVehicleContent.scala
@@ -30,7 +30,7 @@ object SimpleVRPWithVehicleContent extends App{
 
   // Distance
   val vehiclesRouteLength = Array.tabulate(v)(vehicle => CBLSIntVar(m, name = "Route length of vehicle " + vehicle))
-  val routeLengthInvariant = new RouteLength(gc,n,v,vehiclesRouteLength,(from: Long, to: Long) => symmetricDistance(from)(to))
+  val routeLengthInvariant = new RouteLength(gc,n,v,vehiclesRouteLength,(from: Int, to: Int) => symmetricDistance(from)(to))
 
   //Chains
   val precedenceRoute = myVRP.routes.createClone()
@@ -54,10 +54,10 @@ object SimpleVRPWithVehicleContent extends App{
 
   m.close()
 
-  /*def postFilter(node:Long): (Long) => Boolean = {
-    val enoughSpaceAfterNeighborNow: (Long,Long,Array[Long]) => Boolean =
+  /*def postFilter(node:Int): (Int) => Boolean = {
+    val enoughSpaceAfterNeighborNow: (Int,Int,Array[Int]) => Boolean =
       CapacityHelper.enoughSpaceAfterNeighbor(n,capacityInvariant)
-    (neighbor: Long) => {
+    (neighbor: Int) => {
       myVRP.isRouted(neighbor) &&
         enoughSpaceAfterNeighborNow(node,neighbor,contentsFlow)
     }
@@ -70,12 +70,12 @@ object SimpleVRPWithVehicleContent extends App{
   // MOVING
 
   val nextMoveGenerator = {
-    (exploredMoves:List[OnePointMoveMove], t:Option[List[Long]]) => {
-      val chainTail: List[Long] = t match {
+    (exploredMoves:List[OnePointMoveMove], t:Option[List[Int]]) => {
+      val chainTail: List[Int] = t match {
         case None =>
           val movedNode = exploredMoves.head.movedPoint
           chainsExtension.nextNodesInChain(chainsExtension.firstNodeInChainOfNode(movedNode))
-        case Some(tail: List[Long]) => tail
+        case Some(tail: List[Int]) => tail
       }
 
       chainTail match {
@@ -93,7 +93,7 @@ object SimpleVRPWithVehicleContent extends App{
     () => myVRP.routed.value.filter(chainsExtension.isHead),
     ()=> myVRP.kFirst(v*2,closestRelevantPredecessorsByDistance(_)), myVRP,neighborhoodName = "MoveHeadOfChain")
 
-  def lastNodeOfChainMove(lastNode:Long) = onePointMove(
+  def lastNodeOfChainMove(lastNode:Int) = onePointMove(
     () => List(lastNode),
     ()=> myVRP.kFirst(v*2,
       ChainsHelper.relevantNeighborsForLastNodeAfterHead(
@@ -106,7 +106,7 @@ object SimpleVRPWithVehicleContent extends App{
   val oneChainMove = {
     dynAndThen(firstNodeOfChainMove,
       (moveMove: OnePointMoveMove) => {
-        mu[OnePointMoveMove, Option[List[Long]]](
+        mu[OnePointMoveMove, Option[List[Int]]](
           lastNodeOfChainMove(chainsExtension.lastNodeInChainOfNode(moveMove.movedPoint)),
           nextMoveGenerator,
           None,
@@ -115,17 +115,17 @@ object SimpleVRPWithVehicleContent extends App{
       }) name "OneChainMove"
   }
 
-  def onePtMove(k:Long) = profile(onePointMove(myVRP.routed, () => myVRP.kFirst(k,closestRelevantPredecessorsByDistance(_)), myVRP))
+  def onePtMove(k:Int) = profile(onePointMove(myVRP.routed, () => myVRP.kFirst(k,closestRelevantPredecessorsByDistance(_)), myVRP))
 
   // INSERTING
 
   val nextInsertGenerator = {
-    (exploredMoves:List[InsertPointMove], t:Option[List[Long]]) => {
-      val chainTail: List[Long] = t match {
+    (exploredMoves:List[InsertPointMove], t:Option[List[Int]]) => {
+      val chainTail: List[Int] = t match {
         case None =>
           val insertedNode = exploredMoves.head.insertedPoint
           chainsExtension.nextNodesInChain(chainsExtension.firstNodeInChainOfNode(insertedNode))
-        case Some(tail: List[Long]) => tail
+        case Some(tail: List[Int]) => tail
       }
 
       chainTail match {
@@ -143,7 +143,7 @@ object SimpleVRPWithVehicleContent extends App{
     myVRP.kFirst(v*2,closestRelevantPredecessorsByDistance(_))
   }, myVRP,neighborhoodName = "InsertUF")
 
-  def lastNodeOfChainInsertion(lastNode:Long) = insertPointUnroutedFirst(
+  def lastNodeOfChainInsertion(lastNode:Int) = insertPointUnroutedFirst(
     () => List(lastNode),
     ()=> myVRP.kFirst(
       v*2,
@@ -156,7 +156,7 @@ object SimpleVRPWithVehicleContent extends App{
   val oneChainInsert = {
     dynAndThen(firstNodeOfChainInsertion,
       (insertMove: InsertPointMove) => {
-        mu[InsertPointMove,Option[List[Long]]](
+        mu[InsertPointMove,Option[List[Int]]](
           lastNodeOfChainInsertion(chainsExtension.lastNodeInChainOfNode(insertMove.insertedPoint)),
           nextInsertGenerator,
           None,
