+++ oscar-cbls/src/main/scala/oscar/cbls/package.scala
@@ -59,7 +59,7 @@ import scala.language.implicitConversions
   *
   *  //creating variables, one queen par column,
   *  //initialized on a permutation of the diagolal (all on different rows)
-  *  val queens = Array.tabulate(N)((q:Long) => CBLSIntVar(init.next(),0L until N, "queen" + q))
+  *  val queens = Array.tabulate(N)((q:Long) => CBLSIntVar(init.next(),0 until N, "queen" + q))
   *
   *  c.post(allDiff( for (q <- range) yield queens(q) + q) )
   *  c.post(allDiff( for (q <- range) yield q - queens(q)) )
@@ -210,7 +210,7 @@ package object cbls extends ModelingAPI{
 
     def minus(v: SetValue): SetInvariant = Diff(x, v)
 
-    def map(fun: Long => Long,outputDomain:Domain) = SetMap(x, fun, outputDomain)
+    def map(fun: Int => Int,outputDomain:Domain) = SetMap(x, fun, outputDomain)
   }
 
   implicit class IntValueArrayOps(intValueArray: Array[IntValue]) {
@@ -242,7 +242,7 @@ package object cbls extends ModelingAPI{
   }
 
   // implicit conversion of Range towards a RangeHotRestart, to use the StartBy keyword
-  implicit def instrumentRange(r: NumericRange[Long]): InstrumentedRange = new InstrumentedRange(r)
+  implicit def instrumentRange(r: NumericRange[Int]): InstrumentedRange = new InstrumentedRange(r)
 
   //this one has been added followingthe 32 to 64 bits port of oscar.cbls
 /*  implicit def longToInt(value:Long):Int = {
